requirements.txt
anthropic>=0.25.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
PyMySQL>=1.1.0
cryptography>=41.0.0
